aesthetics, etc

+++ Stop-ComputerIn.ps1
@@ -11,7 +11,7 @@ Creates a scheduled task to run a shutdown in 15 minutes and registers it.
 [CmdletBinding()]
 param (
     [Parameter(Position=0)][ValidateRange(1,1439)][int]$Minutes = 10,
-    [Parameter()][Alias("r")][switch]$Restart,
+    [Parameter()][switch]$Restart,
     [Parameter()][switch]$Help
 )
 
@@ -19,27 +19,23 @@ if ($Help) {
     Get-Help $MyInvocation.MyCommand.path; exit
 }
 
-if ($Restart) {
-    $operation = 'restart'
-    $cmdop = 'Restart-Computer'
-} else {
-    $operation = 'shutdown'
-    $cmdop = 'Stop-Computer'
-}
+if ($Restart) { $cmdop = 'Restart-Computer' }
+else          { $cmdop = 'Stop-Computer'    }
+
+$args = "-NoProfile -WindowStyle Hidden -command $cmdop"
 
 # If task exists, remove/overwrite with new one.
 Unregister-ScheduledTask -TaskName PowerShell-ScheduledShutdown `
     -Confirm:$false -ErrorAction SilentlyContinue
 
-Write-Output "Scheduling $operation in $Minutes minutes..."
+$date    = (Get-Date).AddMinutes($Minutes)
+$trigger = New-ScheduledTaskTrigger -Once -At $date
+$action  = New-ScheduledTaskAction -Execute 'powershell.exe' -Argument $args
+$sts     = New-ScheduledTaskSettingsSet -AllowStartIfOnBatteries
+$taskdef = New-ScheduledTask -Action $action -Trigger $trigger -Settings $sts `
+            -Description "Scheduled $cmdop from Powershell"
 
-$date     = (Get-Date).AddMinutes($Minutes)
-$trigger  = New-ScheduledTaskTrigger -Once -At $date
-$action   = New-ScheduledTaskAction -Execute 'powershell.exe' -Argument $cmdop
-$sts      = New-ScheduledTaskSettingsSet -AllowStartIfOnBatteries
-$taskdef  = New-ScheduledTask -Action $action -Trigger $trigger -Settings $sts
-
-Register-ScheduledTask -TaskName PowerShell-ScheduledShutdown `
+$task = Register-ScheduledTask -TaskName PowerShell-ScheduledShutdown `
     -InputObject $taskdef -ErrorAction Inquire
 
-Write-Output "Event scheduled for $date."
+Write-Output "Scheduled $cmdop at $date."
